Adding a json file showing how database arrays should be built:
JSON Object containing the slugs of all tables, built up of an array of Objects:
{
"name"   : The name of the row
"type"   : The datatype of that row
"nn"     : Is it non-null?
"uq pk"  : Is it the Unique Primary Key? (Only once per table)
"ai"     : Is ist auto-increment?
"un"     : Is it unsigned?
"default": The default value